requirements.txt
streamlit
gspread
gspread_dataframe
pandas
google-auth
